[oscar-create] OsmKeyValueObjectStore::addPolyStoreItems: SIGSEGV due to unguarded variable access
Problem:
The Worker used to add the polystore items was not copied but relied on a caching filter.
Thus, updating the filter cache in multiple worker threads resulted in a crash

Solution:
Use thread private workers and move work items into WorkContext and guard it with an atomic bool per work items.